requirements.txt
django>=5.1,<5.2
django-htmx>=1.19
django-environ>=0.11
psycopg2-binary>=2.9
gunicorn>=22.0
whitenoise>=6.7
Pillow>=10.0
